+++ Syns/ISynsStore.cs
@@ -2,6 +2,6 @@ namespace Syns
 {
     public interface ISynsStore
     {
-        int GetTodaySyns(string username);
+        decimal GetTodaySyns(string username);
     }
 }
